op_heads_store: move the ancestry walk into the OpHeadsStore trait
