+++ lib/src/op_heads_store.rs
@@ -12,11 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::collections::HashSet;
 use std::fmt::Debug;
 use std::sync::Arc;
 
 use thiserror::Error;
 
+use crate::dag_walk;
 use crate::op_store::{OpStore, OperationId};
 use crate::operation::Operation;
 
@@ -74,4 +76,20 @@ pub trait OpHeadsStore: Send + Sync + Debug {
     fn lock(&self) -> LockedOpHeads;
 
     fn get_heads(&self, op_store: &Arc<dyn OpStore>) -> Result<OpHeads, OpHeadResolutionError>;
+
+    /// Removes operations in the input that are ancestors of other operations
+    /// in the input. The ancestors are removed both from the list and from
+    /// disk.
+    fn handle_ancestor_ops(&self, op_heads: Vec<Operation>) -> Vec<Operation> {
+        let op_head_ids_before: HashSet<_> = op_heads.iter().map(|op| op.id().clone()).collect();
+        let neighbors_fn = |op: &Operation| op.parents();
+        // Remove ancestors so we don't create merge operation with an operation and its
+        // ancestor
+        let op_heads = dag_walk::heads(op_heads, &neighbors_fn, &|op: &Operation| op.id().clone());
+        let op_head_ids_after: HashSet<_> = op_heads.iter().map(|op| op.id().clone()).collect();
+        for removed_op_head in op_head_ids_before.difference(&op_head_ids_after) {
+            self.remove_op_head(removed_op_head);
+        }
+        op_heads.into_iter().collect()
+    }
 }
